Bug #11148, Patch 1.6.0 to sort recent documents by most recently used
John Hanely:
"Recent documents are shown unsorted, yet limited to a maximum of 25 items.
This means that if you have more than 25 recent documents, only the most recently _added to the list_ will be shown.  That is, if you open a document which has already fallen off the bottom of the list, it will not move to the top of the list.  According to the GTK documentation, recent documents should not be shown unsorted if the count is limited.

I am attaching a patch (against 1.6.0) to sort the list by most recently used."