Avoid inappropriate use of xfce_panel_pixbuf_from_source()
The main purpose of this function is to manage different types of
sources through a unified API, but here we only use icon names.
Moreover, it almost never returns NULL, because it returns a fallback
icon when it doesn't find one from the source given as argument, which
is obviously not the expected behavior here. Finally, it does not allow
to take window scaling properly into account
(xfce_panel_set_image_from_source() could be used instead, but for the
reasons already mentioned, it is better to use a GTK function directly).